heartbeat mechanism between umap_stack and fuzzer
The heartbeat_file's mtime will be updated every time a request from the
victim's host USB stack is received. This allows the fuzzer to monitor the
status of the victim's host USB stack (for example, to identify that the
victim's stack is unresponsive).

Currently going through the filesystem, like the triggering mechanism; however,
depending on the platform, the time-resolution that the mtimes provide may be
limited, so long-term, a different mechanism should probably be used...